refactor: google 로직 분리
