+++ src/main/java/run/attraction/api/v1/auth/service/helper/AuthProviderAndTokenHelper.java
@@ -8,15 +8,15 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
-import run.attraction.api.v1.mail.UserLoggedEvent;
+import run.attraction.api.v1.gmail.event.UserLoggedEvent;
 import run.attraction.api.v1.auth.service.dto.UserTokenDto;
-import run.attraction.api.v1.auth.token.GoogleRefreshToken;
+import run.attraction.api.v1.gmail.entity.GoogleRefreshToken;
 import run.attraction.api.v1.auth.token.LogoutAccessToken;
 import run.attraction.api.v1.auth.token.RefreshToken;
 import run.attraction.api.v1.auth.token.exception.InvalidTokenException;
 import run.attraction.api.v1.auth.token.exception.TokenNotFoundException;
 import run.attraction.api.v1.auth.token.jwt.JwtService;
-import run.attraction.api.v1.auth.token.repository.GoogleRefreshTokenRepository;
+import run.attraction.api.v1.gmail.repository.GoogleRefreshTokenRepository;
 import run.attraction.api.v1.auth.token.repository.LogoutAccessTokenRepository;
 import run.attraction.api.v1.auth.token.repository.RefreshTokenRepository;
 import run.attraction.api.v1.user.User;
