+++ src/main/java/run/attraction/api/v1/gmail/config/KafkaProducerConfig.java
@@ -1,4 +1,4 @@
-package run.attraction.api.v1.mail.config;
+package run.attraction.api.v1.gmail.config;
 
 import java.util.Properties;
 import org.apache.kafka.clients.producer.KafkaProducer;
