+++ src/main/java/run/attraction/api/v1/gmail/repository/GoogleRefreshTokenRepository.java
@@ -1,9 +1,9 @@
-package run.attraction.api.v1.auth.token.repository;
+package run.attraction.api.v1.gmail.repository;
 
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
-import run.attraction.api.v1.auth.token.GoogleRefreshToken;
+import run.attraction.api.v1.gmail.entity.GoogleRefreshToken;
 
 public interface GoogleRefreshTokenRepository extends JpaRepository<GoogleRefreshToken, String> {
   @Query("SELECT rt FROM GoogleRefreshToken rt WHERE rt.email = :userEmail AND rt.shouldReissueToken = true")
